requirements.txt
streamlit
langchain_core
langchain-community
langchain-openai
langchain
chromadb
pypdf
python-dotenv
pdfplumber


